Fixed tests not building.

+++ src/test/java/io/lovelacetech/server/util/ModelUtils.java
@@ -43,7 +43,7 @@ public class ModelUtils {
     user.setId(new UUID(0, index));
     user.setEmail(String.format("test_email_%d@test.com", index));
     user.setPassword(PasswordUtils.encode("test_password_" + index));
-    user.setAccessLevel(AccessLevel.USER.toInt());
+    user.setAccessLevel(AccessLevel.USER);
     user.setCompanyId(new UUID(index, index));
     user.setFirstName("first_name_" + index);
     user.setLastName("last_name_" + index);
